requirements.txt
# Flask==3.0.0
# python-dotenv==1.0.0
# requests==2.31.0

Flask>=3.0.0
python-dotenv>=1.0.1
requests>=2.32.0
openai>=1.51.0
